git-svn-id: https://svn.apache.org/repos/asf/lucene/dev/branches/security@1499601 13f79535-47bb-0310-9956-ffa450edef68

+++ lucene/misc/src/test/org/apache/lucene/misc/SweetSpotSimilarityTest.java
@@ -246,7 +246,7 @@ public class SweetSpotSimilarityTest extends LuceneTestCase {
   
     SweetSpotSimilarity ss = new SweetSpotSimilarity() {
         @Override
-        public float tf(int freq) {
+        public float tf(float freq) {
           return hyperbolicTf(freq);
         }
       };
